Improve documentation, reorganize files, and test (#34)
* work in progress docs

* finish out most of api reference

* linkify

* flesh out more

* wip hello world

* hello world example

* Update getting_started.md

* Update getting_started.md

* Update getting_started.md

* Update getting_started.md

* Update getting_started.md

* reorganize

reorganize file structure

- remove heatmap because we ship one now
- remove funnel because we ship one onw

* Update map resources.md

* map resources aren't relevant here

* examples folder name

* examples readme

* one

* two

* cleanup chord

* fix hello world

* cleanup devices

* cleanup gauge

* cleanup liquid gauge

* smoke tests

* travis

* travis

* badge
