+++ .nvmrc
@@ -1,0 +1,1 @@
+v6.10.2
