+++ .travis.yml
@@ -1,0 +1,2 @@
+language: node_js
+cache:  yarn
