--- docs/three measure heatmap.md
@@ -1,37 +1,0 @@
-
-##Custom Heatmap - with Three Measure Cells##
-
-This heatmap allows you to put three measures into one cell, coloring by a measure of your choice. 
-This is helpful for showing google analytic type measures, as well as more information than a typical heatmap.
-
-![Heatmap image](https://cloud.githubusercontent.com/assets/695747/10524755/28d7decc-734f-11e5-8ad6-da64da12e8eb.png "Heatmap Visualization")
-
-**How it works**
-
-Create a look with one to three measures and two dimensions. The second dimension must be pivotted, The heatmap will make this the column headers. 
-
-**Settings**
-
-
-| Section | Setting Name | Description |
-|---------|--------------|-------------|
-| Data | color range | pre-built or custom color ranges |
-| Data | Custom Color Range | When custom is selected above, put at least two colors to create a custom d3 color range |
-| Data | Min range to color | anything below this range will not be part of the colored cells |
-| Data | Max range to color | Anything above this range will be equal to the maximum color |
-| Data | Measure to color | 1,2, or 3 - indicate the ordinal position of the measure you'd like to use as your color value |
-| Data | Show Cell Borders | Shows cell borders |
-| Data | Show Cell Bordersfor null Values | Shows cell borders around empty cells |
-| Data | Border Color | Changes Cell Border - Default is black |
-| Data | Cell Alignment | Choose top, middle, or bottom alignment |
-| Data | Equal Width Columns | Turns on or off equal width |
-| Heading | Heading Background Color | Choose background color |
-| Heading | Heading Text Color | Choose text color |
-| Heading | Show Borders | Toggles border on/off |
-| Heading | Border Color | Choose border color |
-| Heading | Prefix | Prefix the heading with any string such as "Week # " |
-| Heading | Title | Change the title value |
-
-
-
-**More Images**
