+++ examples/README.md
@@ -1,0 +1,33 @@
+<table>
+    <tr valign="top">
+        <td width="25%">
+        	<a href="hello_world/hello_world.js">
+        	Hello World<br>
+        	<img width="100%" src="hello_world/hello_world.png"></a>
+        </td>
+        <td width="25%">
+        	<a href="chord">
+        	Chord<br>
+        	<img width="100%" src="chord/chord.jpg"></a>
+        </td>
+    </tr>
+     <tr valign="top">
+       <td width="25%">
+        	<a href="single_value_panes/single_value_pane.js">
+        	Single Value Pane<br>
+        	<img width="100%"  src="single_value_panes/pane1.png"></a>
+        </td>
+        <td width="25%">
+        	<a href="devices/devices.js">
+        	Devices<br>
+        	<img width="100%"  src="devices/devices.png"></a>
+        </td>
+    </tr>
+     <tr valign="top">
+       <td width="25%">
+        	<a href="stopwatch/stopwatch.js">
+        	Stopwatch<br>
+        	<img width="100%"  src="stopwatch/stopwatch.png"></a>
+        </td>
+    </tr>
+</table>
