+++ examples/bubbles/bubbles.js
