+++ examples/chord/README.md
@@ -1,0 +1,11 @@
+#  Chord
+
+This diagram allows you to show affinity between two dimensions. This is helpful for showing customer paths, market basket analysis, among others.
+
+![](chord.jpg)
+
+**How it works**
+
+Create a look with one measure and two dimensions. The second dimension must be pivotted, it is better if the dimension with more unique values is not the pivot. The order of the dimensions does not really matter as the color will fall with the higher affinity direction.
+
+For example, in the chord diagram featured above, more flights occur from LAX to ORD compared to ORD to LAX, so the chord color is associated with LAX.
