+++ examples/single_value_panes/Single Value Pane.md
