+++ examples/single_value_panes/pane2.png
