+++ examples/single_value_panes/single_value_panes.js
