+++ examples/small_multiples/small-multiples.js
