+++ examples/stopwatch/stopwatch.js
