+++ examples/stopwatch/stopwatch.png
