--- liquid_fill_gauge.js
@@ -1,483 +1,0 @@
-/*!
- * @license Open source under BSD 2-clause (http://choosealicense.com/licenses/bsd-2-clause/)
- * Copyright (c) 2015, Curtis Bratton
- * All rights reserved.
- *
- * Liquid Fill Gauge v1.1
- */
-
-(function() {
-  function liquidFillGaugeDefaultSettings(){
-      return {
-          showComparison: true,
-          minValue: 0, // The gauge minimum value.
-          maxValue: 100, // The gauge maximum value.
-          circleThickness: 0.05, // The outer circle thickness as a percentage of it's radius.
-          circleFillGap: 0.05, // The size of the gap between the outer circle and wave circle as a percentage of the outer circles radius.
-          circleColor: "#178BCA", // The color of the outer circle.
-          waveHeight: 0.05, // The wave height as a percentage of the radius of the wave circle.
-          waveCount: 1, // The number of full waves per width of the wave circle.
-          waveRiseTime: 1000, // The amount of time in milliseconds for the wave to rise from 0 to it's final height.
-          waveAnimateTime: 18000, // The amount of time in milliseconds for a full wave to enter the wave circle.
-          waveRise: true, // Control if the wave should rise from 0 to it's full height, or start at it's full height.
-          waveHeightScaling: true, // Controls wave size scaling at low and high fill percentages. When true, wave height reaches it's maximum at 50% fill, and minimum at 0% and 100% fill. This helps to prevent the wave from making the wave circle from appear totally full or empty when near it's minimum or maximum fill.
-          waveAnimate: true, // Controls if the wave scrolls or is static.
-          waveColor: "#178BCA", // The color of the fill wave.
-          waveOffset: 0, // The amount to initially offset the wave. 0 = no offset. 1 = offset of one full wave.
-          textVertPosition: .5, // The height at which to display the percentage text withing the wave circle. 0 = bottom, 1 = top.
-          textSize: 1, // The relative height of the text to display in the wave circle. 1 = 50%
-          valueCountUp: true, // If true, the displayed value counts up from 0 to it's final value upon loading. If false, the final value is displayed.
-          displayPercent: true, // If true, a % symbol is displayed after the value.
-          textColor: "#045681", // The color of the value text when the wave does not overlap it.
-          waveTextColor: "#A4DBf8" // The color of the value text when the wave overlaps it.
-      };
-  }
-
-  function loadLiquidFillGauge(elementId, value, config, compareValue) {
-      if(config == null) config = liquidFillGaugeDefaultSettings();
-
-      var gauge = d3.select("#" + elementId);
-      var radius = Math.min(parseInt(gauge.style("width")), parseInt(gauge.style("height")))/2;
-      var locationX = parseInt(gauge.style("width"))/2 - radius;
-      var locationY = parseInt(gauge.style("height"))/2 - radius;
-      if (config.showComparison && compareValue && !isNaN(compareValue))
-        config.maxValue = compareValue
-      var fillPercent = Math.max(config.minValue, Math.min(config.maxValue, value))/config.maxValue;
-
-      var waveHeightScale;
-      if(config.waveHeightScaling){
-          waveHeightScale = d3.scale.linear()
-              .range([0,config.waveHeight,0])
-              .domain([0,50,100]);
-      } else {
-          waveHeightScale = d3.scale.linear()
-              .range([config.waveHeight,config.waveHeight])
-              .domain([0,100]);
-      }
-
-      var textPixels = (config.textSize*radius/2);
-      var textFinalValue = parseFloat(value).toFixed(2);
-      var textStartValue = config.valueCountUp?config.minValue:textFinalValue;
-      var percentText = config.displayPercent?"%":"";
-      var circleThickness = config.circleThickness * radius;
-      var circleFillGap = config.circleFillGap * radius;
-      var fillCircleMargin = circleThickness + circleFillGap;
-      var fillCircleRadius = radius - fillCircleMargin;
-      var waveHeight = fillCircleRadius*waveHeightScale(fillPercent*100);
-
-      var waveLength = fillCircleRadius*2/config.waveCount;
-      var waveClipCount = 1+config.waveCount;
-      var waveClipWidth = waveLength*waveClipCount;
-
-      // Rounding functions so that the correct number of decimal places is always displayed as the value counts up.
-      var textRounder = function(value){ return Math.round(value); };
-      if(parseFloat(textFinalValue) != parseFloat(textRounder(textFinalValue))){
-          textRounder = function(value){ return parseFloat(value).toFixed(1); };
-      }
-      if(parseFloat(textFinalValue) != parseFloat(textRounder(textFinalValue))){
-          textRounder = function(value){ return parseFloat(value).toFixed(2); };
-      }
-
-      // Data for building the clip wave area.
-      var data = [];
-      for(var i = 0; i <= 40*waveClipCount; i++){
-          data.push({x: i/(40*waveClipCount), y: (i/(40))});
-      }
-
-      // Scales for drawing the outer circle.
-      var gaugeCircleX = d3.scale.linear().range([0,2*Math.PI]).domain([0,1]);
-      var gaugeCircleY = d3.scale.linear().range([0,radius]).domain([0,radius]);
-
-      // Scales for controlling the size of the clipping path.
-      var waveScaleX = d3.scale.linear().range([0,waveClipWidth]).domain([0,1]);
-      var waveScaleY = d3.scale.linear().range([0,waveHeight]).domain([0,1]);
-
-      // Scales for controlling the position of the clipping path.
-      var waveRiseScale = d3.scale.linear()
-          // The clipping area size is the height of the fill circle + the wave height, so we position the clip wave
-          // such that the it will overlap the fill circle at all when at 0%, and will totally cover the fill
-          // circle at 100%.
-          .range([(fillCircleMargin+fillCircleRadius*2+waveHeight),(fillCircleMargin-waveHeight)])
-          .domain([0,1]);
-      var waveAnimateScale = d3.scale.linear()
-          .range([0, waveClipWidth-fillCircleRadius*2]) // Push the clip area one full wave then snap back.
-          .domain([0,1]);
-
-      // Scale for controlling the position of the text within the gauge.
-      var textRiseScaleY = d3.scale.linear()
-          .range([fillCircleMargin+fillCircleRadius*2,(fillCircleMargin+textPixels*0.7)])
-          .domain([0,1]);
-
-      // Center the gauge within the parent SVG.
-      var gaugeGroup = gauge.append("g")
-          .attr('transform','translate('+locationX+','+locationY+')');
-
-      // Draw the outer circle.
-      var gaugeCircleArc = d3.svg.arc()
-          .startAngle(gaugeCircleX(0))
-          .endAngle(gaugeCircleX(1))
-          .outerRadius(gaugeCircleY(radius))
-          .innerRadius(gaugeCircleY(radius-circleThickness));
-      gaugeGroup.append("path")
-          .attr("d", gaugeCircleArc)
-          .style("fill", config.circleColor)
-          .attr('transform','translate('+radius+','+radius+')');
-
-      // Text where the wave does not overlap.
-      var text1 = gaugeGroup.append("text")
-          .text(textRounder(textStartValue) + percentText)
-          .attr("class", "liquidFillGaugeText")
-          .attr("text-anchor", "middle")
-          .attr("font-size", textPixels + "px")
-          .style("fill", config.textColor)
-          .attr('transform','translate('+radius+','+textRiseScaleY(config.textVertPosition)+')');
-
-      // The clipping wave area.
-      var clipArea = d3.svg.area()
-          .x(function(d) { return waveScaleX(d.x); } )
-          .y0(function(d) { return waveScaleY(Math.sin(Math.PI*2*config.waveOffset*-1 + Math.PI*2*(1-config.waveCount) + d.y*2*Math.PI));} )
-          .y1(function(d) { return (fillCircleRadius*2 + waveHeight); } );
-      var waveGroup = gaugeGroup.append("defs")
-          .append("clipPath")
-          .attr("id", "clipWave" + elementId);
-      var wave = waveGroup.append("path")
-          .datum(data)
-          .attr("d", clipArea)
-          .attr("T", 0);
-
-      // The inner circle with the clipping wave attached.
-      var fillCircleGroup = gaugeGroup.append("g")
-          .attr("clip-path", "url(#clipWave" + elementId + ")");
-      fillCircleGroup.append("circle")
-          .attr("cx", radius)
-          .attr("cy", radius)
-          .attr("r", fillCircleRadius)
-          .style("fill", config.waveColor);
-
-      // Text where the wave does overlap.
-      var text2 = fillCircleGroup.append("text")
-          .text(textRounder(textStartValue) + percentText)
-          .attr("class", "liquidFillGaugeText")
-          .attr("text-anchor", "middle")
-          .attr("font-size", textPixels + "px")
-          .style("fill", config.waveTextColor)
-          .attr('transform','translate('+radius+','+textRiseScaleY(config.textVertPosition)+')');
-
-      // Make the value count up.
-      if(config.valueCountUp){
-          var textTween = function(){
-              var i = d3.interpolate(this.textContent, textFinalValue);
-              return function(t) { this.textContent = textRounder(i(t)) + percentText; }
-          };
-          text1.transition()
-              .duration(config.waveRiseTime)
-              .tween("text", textTween);
-          text2.transition()
-              .duration(config.waveRiseTime)
-              .tween("text", textTween);
-      }
-
-      // Make the wave rise. wave and waveGroup are separate so that horizontal and vertical movement can be controlled independently.
-      var waveGroupXPosition = fillCircleMargin+fillCircleRadius*2-waveClipWidth;
-      if(config.waveRise){
-          waveGroup.attr('transform','translate('+waveGroupXPosition+','+waveRiseScale(0)+')')
-              .transition()
-              .duration(config.waveRiseTime)
-              .attr('transform','translate('+waveGroupXPosition+','+waveRiseScale(fillPercent)+')')
-              .each("start", function(){ wave.attr('transform','translate(1,0)'); }); // This transform is necessary to get the clip wave positioned correctly when waveRise=true and waveAnimate=false. The wave will not position correctly without this, but it's not clear why this is actually necessary.
-      } else {
-          waveGroup.attr('transform','translate('+waveGroupXPosition+','+waveRiseScale(fillPercent)+')');
-      }
-
-      if(config.waveAnimate) animateWave();
-
-      function animateWave() {
-          wave.attr('transform','translate('+waveAnimateScale(wave.attr('T'))+',0)');
-          wave.transition()
-              .duration(config.waveAnimateTime * (1-wave.attr('T')))
-              .ease('linear')
-              .attr('transform','translate('+waveAnimateScale(1)+',0)')
-              .attr('T', 1)
-              .each('end', function(){
-                  wave.attr('T', 0);
-                  animateWave(config.waveAnimateTime);
-              });
-      }
-
-      function GaugeUpdater(){
-          this.update = function(value){
-              var newFinalValue = parseFloat(value).toFixed(2);
-              var textRounderUpdater = function(value){ return Math.round(value); };
-              if(parseFloat(newFinalValue) != parseFloat(textRounderUpdater(newFinalValue))){
-                  textRounderUpdater = function(value){ return parseFloat(value).toFixed(1); };
-              }
-              if(parseFloat(newFinalValue) != parseFloat(textRounderUpdater(newFinalValue))){
-                  textRounderUpdater = function(value){ return parseFloat(value).toFixed(2); };
-              }
-
-              var textTween = function(){
-                  var i = d3.interpolate(this.textContent, parseFloat(value).toFixed(2));
-                  return function(t) { this.textContent = textRounderUpdater(i(t)) + percentText; }
-              };
-
-              text1.transition()
-                  .duration(config.waveRiseTime)
-                  .tween("text", textTween);
-              text2.transition()
-                  .duration(config.waveRiseTime)
-                  .tween("text", textTween);
-
-              var fillPercent = Math.max(config.minValue, Math.min(config.maxValue, value))/config.maxValue;
-              var waveHeight = fillCircleRadius*waveHeightScale(fillPercent*100);
-              var waveRiseScale = d3.scale.linear()
-                  // The clipping area size is the height of the fill circle + the wave height, so we position the clip wave
-                  // such that the it will overlap the fill circle at all when at 0%, and will totally cover the fill
-                  // circle at 100%.
-                  .range([(fillCircleMargin+fillCircleRadius*2+waveHeight),(fillCircleMargin-waveHeight)])
-                  .domain([0,1]);
-              var newHeight = waveRiseScale(fillPercent);
-              var waveScaleX = d3.scale.linear().range([0,waveClipWidth]).domain([0,1]);
-              var waveScaleY = d3.scale.linear().range([0,waveHeight]).domain([0,1]);
-              var newClipArea;
-              if(config.waveHeightScaling){
-                  newClipArea = d3.svg.area()
-                      .x(function(d) { return waveScaleX(d.x); } )
-                      .y0(function(d) { return waveScaleY(Math.sin(Math.PI*2*config.waveOffset*-1 + Math.PI*2*(1-config.waveCount) + d.y*2*Math.PI));} )
-                      .y1(function(d) { return (fillCircleRadius*2 + waveHeight); } );
-              } else {
-                  newClipArea = clipArea;
-              }
-
-              var newWavePosition = config.waveAnimate?waveAnimateScale(1):0;
-              wave.transition()
-                  .duration(0)
-                  .transition()
-                  .duration(config.waveAnimate?(config.waveAnimateTime * (1-wave.attr('T'))):(config.waveRiseTime))
-                  .ease('linear')
-                  .attr('d', newClipArea)
-                  .attr('transform','translate('+newWavePosition+',0)')
-                  .attr('T','1')
-                  .each("end", function(){
-                      if(config.waveAnimate){
-                          wave.attr('transform','translate('+waveAnimateScale(0)+',0)');
-                          animateWave(config.waveAnimateTime);
-                      }
-                  });
-              waveGroup.transition()
-                  .duration(config.waveRiseTime)
-                  .attr('transform','translate('+waveGroupXPosition+','+newHeight+')')
-          }
-      }
-
-      return new GaugeUpdater();
-  }
-
-  var viz = {
-    id: 'liquid_fill_gauge',
-    label: 'Liquid Fill Gauge',
-    options: {
-      showComparison: {
-        label: "Use field comparison",
-        default: false,
-        section: "Value",
-        type: "boolean"
-      },
-      maxValue: {
-        label: "Max value",
-        min: 0,
-        default: 100,
-        section: "Value",
-        type: "number",
-        placeholder: "Any positive number",
-        hidden: function(options) {
-          return options.showComparison
-        }
-      },
-      minValue: {
-        label: "Min value",
-        min: 0,
-        default: 0,
-        section: "Value",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      circleThickness: {
-        label: "Circle thickness",
-        min: 0,
-        default: 0.05,
-        section: "Style",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      circleGap: {
-        label: "Circle gap",
-        min: 0,
-        default: 0.05,
-        section: "Style",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      circleColor: {
-        label: "Circle Color",
-        default: "#8B7DA8",
-        section: "Style",
-        type: "string",
-        display: "color"
-      },
-      waveHeight: {
-        label: "Wave Height",
-        min: 0,
-        default: 0.05,
-        section: "Waves",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      waveCount: {
-        label: "Wave Count",
-        min: 0,
-        max: 100,
-        default: 1,
-        section: "Waves",
-        type: "number",
-        placeholder: "0-100"
-      },
-      waveRiseTime: {
-        label: "Wave Rise Time",
-        min: 0,
-        default: 1000,
-        section: "Waves",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      waveAnimateTime: {
-        label: "Wave Animation Time",
-        min: 0,
-        default: 1800,
-        section: "Waves",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      waveRise: {
-        label: "Wave Rise from 0",
-        default: true,
-        section: "Waves",
-        type: "boolean"
-      },
-      waveHeightScaling: {
-        label: "Scale waves if high or low",
-        default: true,
-        section: "Waves",
-        type: "boolean"
-      },
-      waveAnimate: {
-        label: "Animate Waves",
-        default: true,
-        section: "Waves",
-        type: "boolean"
-      },
-      waveColor: {
-        label: "Wave Color",
-        default: "#64518A",
-        section: "Style",
-        type: "string",
-        display: "color"
-      },
-      waveOffset: {
-        label: "Wave Offset",
-        min: 0,
-        default: 0,
-        section: "Waves",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      textVertPosition: {
-        label: "Text Vertical Offset",
-        min: 0,
-        default: .5,
-        section: "Value",
-        type: "number",
-        placeholder: "Any positive number"
-      },
-      textSize: {
-        label: "Text Size",
-        min: 0,
-        default: 1,
-        section: "Value",
-        type: "number",
-        placeholder: "Number relative to circle size"
-      },
-      valueCountUp: {
-        label: "Animate to Value",
-        default: true,
-        section: "Waves",
-        type: "boolean"
-      },
-      displayPercent: {
-        label: "Display as Percent",
-        default: true,
-        section: "Value",
-        type: "boolean"
-      },
-      textColor: {
-        label: "Text Color (non-overlapped)",
-        default: "#000000",
-        section: "Style",
-        type: "string",
-        display: "color"
-      },
-      waveTextColor: {
-        label: "Text Color (overlapped)",
-        default: "#FFFFFF",
-        section: "Style",
-        type: "string",
-        display: "color"
-      }
-    },
-
-    create: function(element, settings) {
-      this.id = _.uniqueId("fill-gauge-");
-      $elem = d3.select(element).append("svg");
-      $elem.attr("id", this.id);
-    },
-
-    update: function(data, element, settings, resp) {
-      config = liquidFillGaugeDefaultSettings();
-      config = _.extend(config, settings);
-      var compareField = null
-      var compareCell = {value: null}
-      if (resp.fields.measure_like.length > 0)
-        datumField = resp.fields.measure_like[0]
-      else if (resp.fields.dimension_like.length > 0)
-        datumField = resp.fields.dimension_like[0]
-      if (datumField)
-        datum = data[0][datumField.name]
-
-      if (config.showComparison && data)
-        if (resp.fields.measure_like.length > 1)
-          compareField = resp.fields.measure_like[1]
-        else if (resp.fields.dimension_like.length > 1 && resp.fields.measure_like.length == 0)
-          compareField = resp.fields.dimension_like[1]
-        else if (resp.fields.measure_like.length == 1 && resp.fields.dimension_like.length >= 1)
-          compareField = resp.fields.dimension_like[0]
-        else
-          compareField = resp.fields.measure_like[0] || resp.fields.dimension_like[0]
-
-        if (compareField)
-          compareCell = data[0][compareField.name]
-
-        if (config.displayPercent && config.showComparison && compareCell.value > 0)
-          datum.value = datum.value / compareCell.value * 100
-
-      $elem = d3.select(element).selectAll("svg#" + this.id).data([datum])
-      $elem.enter()
-      $elem.html("")
-      $elem.attr("width", "100%")
-      $elem.attr("height", element.clientHeight)
-      $elem.exit().remove()
-      this.gauge = loadLiquidFillGauge(this.id, datum.value, config, compareCell.value);
-
-    }
-  };
-  looker.plugins.visualizations.add(viz);
-
-}());
