--- map resources.md
@@ -1,14 +1,0 @@
-
-Looker uses topoJSON maps for its visualizations. There is a few helpful discourse articles on making your own as well as resources out there of working maps:
-
-###Discourse Articles###
-- </https://discourse.looker.com/t/turn-a-shapefile-into-a-custom-topojson-to-visualize-a-map-in-looker/480>
-- </https://discourse.looker.com/t/using-custom-topojson-maps-in-looker/138>
-- </https://discourse.looker.com/t/creating-custom-maps-files-from-shapefiles/196>
-
-###Existing Maps###
-
-- US Maps by zip code: <https://github.com/ericfeinstein/maps/>
-- New Jersey Counties: <https://gist.github.com/mirez/6468870/>
-
-###Other Resources###
