+++ test/looker_stub.js
@@ -1,0 +1,14 @@
+var allVisualizations = [];
+
+module.exports = {
+  plugins: {
+    visualizations: {
+      add: function(vis) {
+        allVisualizations.push(vis);
+      },
+      all: function() {
+        return allVisualizations;
+      }
+    }
+  }
+}
